+++ entity/src/epilot/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.4'
+    sdk_version: str = '2.1.5'
     gen_version: str = '2.173.0'
-    user_agent: str = 'speakeasy-sdk/python 2.1.4 2.173.0 1.0.0 epilot-entity'
+    user_agent: str = 'speakeasy-sdk/python 2.1.5 2.173.0 1.0.0 epilot-entity'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: